server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
rectpack==0.2.2
pydantic==2.9.2
shapely==2.0.6
reportlab==4.2.5
openpyxl==3.1.5
argon2-cffi==23.1.0
PyJWT==2.9.0